XiaozhiServerPlus/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
openai